resolved conflcts
